Additional performance improvements (#29)
* Slightly improve performance of guarded mode and add benchmark to prove it.

Its faster to iterate over an array than over dictionary keys. This is even true if you have to convert the dictionary keys to an array before.

* Improve performance of expressions which are or can be folded to a constant.

If the AST is a single constant, we can return it directly as lambda.

Also adds a benchmark to prove it and moves all comparing benchmarks to their own dir.

* WIP: Actually disable cache and check performance impact

Slower if re-evaluating with evaluate. Faster if evaluating every expression once. Inconclusive if re-evaluating using delegate.

* Add benchmark cases and improve structure.

* Move mode-check in interpreter from evaluation to build phase.

This helps minimizing the overhead of the checks, since we only build once, but we evaluate often. In worst case, this checks as often as the previous solution, so no negative performance impact is to be expected.

This solution is consistent with the approach taken in the DynamicCompiler.

* Fix wrong executor type