Pop dynamic function stack after every function. Fixes #42.
We're counting arguments in case of dynamic functions, since they can have an arbitrary amount of arguments. Since the count is pretty much as cheap as repeated checks if we're handling a dynamic function, we're always counting, but we only use the result in case of a dynamic function.

Unfortunately, we were missing the pop in case of a static function, so if we have a static function nested inside of a dynamic function, we were looking at the wrong value.

This change always pops the count and thus fixes this issue.

We could think about introducing a dedicated function type for dynamic functions and separating the handling of dynamic and static functions. But this might also introduce more complexity and/or duplication into the code, so we should consider this carefully.

This is not an ugly fix, so we can accept it as solution for now.