Maintenance/net version (#58)
* Update target frameworks as discussed in #54

* Update readme and workflow files to reflect .NET upgrade

* Remove .NET 8 from direct targets since there is no real world benefit at the moment

* ADR: there is no technical benefit in doing this, since there is no real dependency
on .net6 (which reached EOL) here. the target framework only specifies
the language level available to the project and can be read as minimum
requirement. however, the specifics of the .net build system aren't
obvious to many developers and so this has been the cause of confusion
already twice. i hope this commit prevents this problem in the future
even though the only thing this technically achieves is making the
build for .net6 projects fall back to .netstandard2.0...